Move "model" namespaces to their own dirs
