--- apps/web/modules/models/Name.ts
@@ -1,1 +1,0 @@
-export {};
