+++ apps/web/modules/services/import.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from 'fs';
 import path from 'path';
 import { v4 as uuid } from 'uuid';
 import { describe, expect, it } from 'vitest';
-import { Triple } from '../models/Triple';
+import { Triple } from '../triple';
 import { convertLegacyHealthData, eavRowsToTriples, importCSVFile, readFileAsText, unique } from './import';
 
 function readMockFile(filename: string) {
